pyproject.toml
[project]
name = "code-dev-assistant"
version = "0.1.0"
description = "Localized Code Development and Git Assistant with RAG and LLM integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9.1",
    "gitpython>=3.1.40",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "lancedb>=0.16.0",
    "sentence-transformers>=3.0.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "autogen-agentchat>=0.4.0",
    "flask>=3.0.0",
]
[[project.authors]]
name = "nkitan"
email = "ankitdas2k@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mcp>=1.9.1",
]

[project.scripts]
code-dev-assistant = "code_dev_assistant:main"
